scheduler: task.Data made 64bit to avoid overflow

+++ src/runtime/scheduler.go
@@ -88,7 +88,7 @@ func addSleepTask(t *task.Task, duration timeUnit) {
 			panic("runtime: addSleepTask: expected next task to be nil")
 		}
 	}
-	t.Data = uint(duration) // TODO: longer durations
+	t.Data = uint64(duration)
 	now := ticks()
 	if sleepQueue == nil {
 		scheduleLog("  -> sleep new queue")
